Move weekly routes into weekly controller
